2.3-5 Solution
First off, I over-engineered this algorithm by using arrays instead of indices as inputs. Secondly, the determination of the running time failed when I tried to apply the equation from earlier problems. I assume it does not apply and the intent is to intuitively come up with the halving solution.